db: normalize schema (phases 0/1, 3, 5a) + collapse stock footgun (5b)
Schema normalization pass against the dev Supabase project, with matching
Drift (v45->v46) + client changes. No data/users yet, so changes are direct.

Phase 0/1 — drop genuinely-redundant objects:
  - inventory_movements (duplicate of richer stock_ledger), sync_queue
    (abandoned outbox; per-row sync_status is used instead)
  - 2 duplicate indexes (kept the unique/constraint-backed variants)

Phase 3 — expenses (single snapshot model): drop the never-populated FK
  columns (expense_category_id, employee_id, approved_by) and the unused
  expense_categories table. The app already used the denormalized snapshot.

Phase 5a — add business_id + created_at to transaction_items / refund_items
  (additive, backfilled) for direct tenant-scoped RLS/analytics.

Phase 5b — eliminate the dual int/decimal "fractional footgun": stock and
  inventory_levels.quantity become a single decimal column end to end.
  Removed the int/decimal branching in StockMovementService /
  recipe_consumption_service and propagated double through entities, repos,
  cubits, models, widgets, sync, and the remote DS. Added qtyLabel() so unit
  stock still renders without a trailing ".0".

Deferred by decision: Phase 4 (permissions/identity consolidation) and the
uuid+FK "unification" on procurement/recipes/sales-items (the missing FKs are
an intentional snapshot design documented in the table classes).

Verified: flutter analyze clean; all 41 tests pass.

+++ lib/features/inventory/data/inventory_data.dart
@@ -20,11 +20,12 @@ class InventoryItem extends Equatable {
   final String? sku;
 
   /// Stock keyed by branchId. A null key means "global / no branch".
-  final Map<String, int> stockByBranch;
+  /// Decimal — holds whole units and fractional quantities alike.
+  final Map<String, double> stockByBranch;
 
   /// Total stock across all branches (or from product_variants.stock when
-  /// no per-branch entries exist yet).
-  final int totalStock;
+  /// no per-branch entries exist yet). Decimal (unit + fractional products).
+  final double totalStock;
 
   /// Reorder threshold from product_variants.lowStockAlert (0 when unset).
   final int reorderLevel;
